update travis.yml
we don't test "cabal install". we test building with stack only. Plus,
we don't test wild-bind-x11 and others depending on it for most
cases. This is because wild-bind-x11 requires fold-debounce package,
which is in LTSHaskell in lts-6. Plus, we don't test wild-bind-x11 and
others in OSX. I'm not sure X11 builds fine in OSX.